Add listener for follow,like to action
