+++ src/main/java/com/recommendation/data_consumption/action/FollowListener.java
@@ -1,24 +1,25 @@
-package com.recommendation.data_consumption.state;
+package com.recommendation.data_consumption.action;
 
-import com.recommendation.data_consumption.dto.LikeKafkaMessage;
+import com.recommendation.data_consumption.dto.FollowKafkaMessage;
 import com.recommendation.data_consumption.dto.SubscribeContestKafkaMessage;
 import com.recommendation.data_consumption.dto.UpdateMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.core.KafkaTemplate;
 
-public class LikeListener {
+public class FollowListener {
+
     @Autowired
     KafkaTemplate<String,UpdateMessage> updateKafkaTemplate;
 
-    @KafkaListener(topics = "LIKE", containerFactory = "likeKafkaListenerFactory")
-    public void processSubscribeMessage(LikeKafkaMessage likeKafkaMessage)
+    @KafkaListener(topics = "FOLLOW", containerFactory = "followKafkaListenerFactory")
+    public void processSubscribeMessage(FollowKafkaMessage followKafkaMessage)
     {
         UpdateMessage updateMessage=new UpdateMessage();
         updateMessage.setUpdateUnit("POINTS");
         updateMessage.setUpdateValue(1);
-        updateMessage.setRowId(likeKafkaMessage.getUserId());
-        updateMessage.setColumnId(likeKafkaMessage.getCategory());
+        updateMessage.setRowId(followKafkaMessage.getUserId());
+        updateMessage.setColumnId(followKafkaMessage.getUserIdForFollowed());
         updateMessage.setTarget("SURGE");
 
         updateKafkaTemplate.send("UPDATE",updateMessage);
